pyproject.toml
[project]
name = "mirror"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pydantic-settings>=2.9.1",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.45.1",
    "taskiq>=0.11.17",
    "taskiq-aio-pika>=0.4.2",
    "taskiq-redis>=1.0.9",
]
